Fixing Zip mode camera for Airdrop and Interference
If you are using Zip mode, the default camera for "Specialist_CombatProtocol" shows a close-up of your Specialist,  but by the time the camera moves away, the animation is already finished. With no CinescriptCameraType set the camera movement looks fine on Zip mode. For non-Zip mode I did not see noticeable difference. Therefore, I believe it is better to not have CinescriptCameraType set at all, unless there is some unforeseen consequence of changing it which I am not aware of.