Balancing SupplyCost/TradingPostValue for some items
Currently UIBlackMarket_SellItem.PopulateData ensures that the sell price of the item is not more than its cost from engineering. For the items with the same SupplyCost and TradingPostValue it leads to the situation, where if they were chosen as an "Interest" item by black market its selling cost would not change and remain the same as before  Which in turn leads to the feeling of being ripped off of selling opportunity for this month and incentivizes tedious practices (like, hiding scopes specifically in random weapons just before months roll over).

While it may be argued that it is working as intended and the problem lies solely in managing expectations, I believe it is better to standartize having TradingPostValue lower than SupplyCost, so that the "black market being interested in" for these items would actually do something. The solution for such items would be to either lower TradingPostValue or raise SupplyCost. Generally, I think it is easier to use the former change, as many of the items with such entries are rarely sold anyways. For the scopes. specifically, given that there is precedent for weapon upgrade with TradingPostValue of 10 (Suppressors) I chose the latter approach.

Note, that for many items "interested in" price would still be lower than usually expected x2, which is arguably unnecessary confusing. However, changing this would either require more sizable change to the pricing values, or the change to theUIBlackMarket_SellItem. It seems to me that X2CommunityCore#859 may help with that in the future, but it is currently pending.